requirements.txt
altgraph==0.17
anytree==2.8.0
fbs==0.9.0
future==0.18.2
ImageHash==4.1.0
macholib==1.14
numpy==1.19.1
pefile==2019.4.18
Pillow==7.2.0
PyInstaller==3.4
PyQt5==5.15.0
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
